requirements.txt
aiocsv==1.3.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp_socks==0.10.1
aiolimiter==1.2.1
aiosignal==1.3.1
aiosqlite==0.21.0
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.8.0
APScheduler==3.11.0
art==6.4
attrs==24.2.0
better-proxy==1.3.1
blessed==1.20.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cloudscraper==1.2.71
colorama==0.4.6
curl_cffi==0.7.4
editor==1.6.6
fake-useragent==1.5.1
Faker==30.3.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imap-tools==1.10.0
inquirer==3.3.0
iso8601==2.1.0
jinxed==1.3.0
loguru==0.7.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
names==0.3.0
numpy==1.26.4
outcome==1.3.0.post0
package_name==0.1
propcache==0.2.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyfiglet==1.0.2
Pygments==2.19.1
pyparsing==3.2.1
pypika-tortoise==0.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-socks==2.7.1
python-telegram-bot==21.11.1
pytz==2024.2
PyYAML==6.0.2
readchar==4.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
runs==1.2.2
selenium==4.29.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tornado==6.4.2
tortoise-orm==0.24.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
undetected-chromedriver==3.5.5
urllib3==2.2.3
user_agent==0.1.10
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0
win32_setctime==1.2.0
wsproto==1.2.0
xmod==1.8.1
yarl==1.14.0
